аватар полностью, новая карточка лайк со 2 раза, ux работает. нет промисов и проверить ошибку удаления карточки

+++ src/components/cards.js
@@ -1,6 +1,6 @@
 import { config } from "./api";
 import { render } from "./index";
-import { likeNum } from "./card";
+// import { likeNum } from "./card";
 
 export const getInitialCards = () => {
   return fetch(`${config.baseUrl}/cards`, {
@@ -12,45 +12,32 @@ export const getInitialCards = () => {
       }
     })
     .then((res) => {
-     console.log(res);
+      console.log(res);
       const initialCards = Array.from(res);
-       
-      render(initialCards); 
-  
+
+      render(initialCards);
     })
-   
+
     .catch((res) => {
       return Promise.reject(`Ошибка: ${res.status}`);
-      
     });
-    
 };
 export function addNewCard(name, link, likeNum) {
-  let userId = '18224dc979a1237fbf3f98ed';
-  fetch("https://nomoreparties.co/v1/wff-cohort-22/cards", {
+  let userId = "18224dc979a1237fbf3f98ed";
+  fetch(`${config.baseUrl}/cards`, {
     method: "POST",
-    headers: {
-      authorization: "3bdecd97-cc83-4e5e-ac8d-e22694049ffd",
-      "Content-Type": "application/json",
-    },
+    headers: config.headers,
     body: JSON.stringify({
       name: name,
       link: link,
       likes: likeNum,
-     
     }),
-  })
-    .then((res) => res.json())
-  
-  
+  }).then((res) => res.json());
 }
 export function deleteCard(id, listItem) {
   fetch(`${config.baseUrl}/cards/${id}`, {
     method: "DELETE",
-    headers: {
-      authorization: "3bdecd97-cc83-4e5e-ac8d-e22694049ffd",
-      "Content-Type": "application/json",
-    },
+    headers: config.headers,
   })
     .then((response) => {
       return response.json();
@@ -59,67 +46,48 @@ export function deleteCard(id, listItem) {
   listItem.remove();
 }
 export function likeCard(item, likeButton, id, userId) {
-getInitialCards();
-  // 
-  //likeButton.nextElementSibling.textContent -= likeButton.classList.toggle("card__like-button_is-active" ? -1: 1);
-let numLike = item.likes.length; 
-if (numLike == 0) {addLikeCard( id);
- likeButton.nextElementSibling.textContent = numLike + 1; 
- likeButton.classList.add("card__like-button_is-active");
-}
-else {
- const lk = Array.from(item.likes);
- const likes = lk.map((elem) => ({ _id: elem._id}));
-for (let el in likes)
- {if (likes[el]._id== userId){ 
-  deleteLikeCard(id);
-  //  likeButton.nextElementSibling.textContent = numLike - 1; 
-   likeButton.classList.remove("card__like-button_is-active");
-  } else {addLikeCard(id);
-    // likeButton.nextElementSibling.textContent = numLike + 1; 
+  getInitialCards();
+
+  let numLike = item.likes.length;
+  if (numLike == 0) {
+    addLikeCard(id);
+    likeButton.nextElementSibling.textContent = numLike + 1;
     likeButton.classList.add("card__like-button_is-active");
-}}
+  } else {
+    const arrayLikes = Array.from(item.likes);
+    const likes = arrayLikes.map((elem) => ({ _id: elem._id }));
+    for (let el in likes) {
+      if (likes[el]._id == userId) {
+        deleteLikeCard(id);
+        //  likeButton.nextElementSibling.textContent = numLike - 1;
+        likeButton.classList.remove("card__like-button_is-active");
+      } else {
+        addLikeCard(id);
+        // likeButton.nextElementSibling.textContent = numLike + 1;
+        likeButton.classList.add("card__like-button_is-active");
+      }
+    }
+  }
+
+  getInitialCards();
 }
-    
 
-
-getInitialCards();
-}
-
-
-export function addLikeCard( id) {
-  
+export function addLikeCard(id) {
   fetch(`${config.baseUrl}/cards/likes/${id}`, {
     method: "PUT",
-    headers: {
-      authorization: "3bdecd97-cc83-4e5e-ac8d-e22694049ffd",
-      "Content-Type": "application/json",
-    },
-  })
-    .then((response) => {
-      return response.json();
-    })
+    headers: config.headers,
     
-       
-     // if (((res.likes[0]._id) !== userId))
-   
-   
-   
-    
-  
+  }).then((response) => {
+    return response.json();
+  });
+
+  // if (((res.likes[0]._id) !== userId))
 }
 export function deleteLikeCard(id) {
   fetch(`${config.baseUrl}/cards/likes/${id}`, {
     method: "DELETE",
-    headers: {
-      authorization: "3bdecd97-cc83-4e5e-ac8d-e22694049ffd",
-      "Content-Type": "application/json",
-    },
-  })
-    .then((response) => {
-      return response.json();
-    })
-    
-    
-  
+    headers: config.headers,
+  }).then((response) => {
+    return response.json();
+  });
 }
